% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_z_project.R
\name{grouped_z_project}
\alias{grouped_z_project}
\title{grouped_z_project
merged images}
\usage{
grouped_z_project(im, grouped_size, FUN = sum)
}
\arguments{
\item{im}{\link{array} (\strong{optional}) an image as matrix or array}

\item{grouped_size}{\link{numeric} (\strong{with default}) number of image merged in a group}

\item{FUN}{\link{function}  (\strong{with default})  function between the merged images (default FUN=sum)}
}
\value{
tmp, n \link{list}
}
\description{
grouped_z_project
merged images
}
\examples{
 \dontrun{
tmp<-grouped_z_project(im,85)
}
}
