% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-noise.R
\name{local_median}
\alias{local_median}
\title{local_median
calcul of median with size radius x radius}
\usage{
local_median(image, radius, bg = 0)
}
\value{
med_mat
}
\description{
local_median
calcul of median with size radius x radius
}
