% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-noise.R
\name{remove_noise}
\alias{remove_noise}
\title{remove-noise
remove noise by deleting outliers}
\usage{
remove_noise(im, radius = 5, threshold = 50, bg = 0)
}
\arguments{
\item{im}{\link{array} (\strong{optional}) an image as matrix or array}

\item{radius}{\link{numeric} (\strong{with default}) radius value (default radius=5)}

\item{threshold}{\link{numeric}  (\strong{with default})  threshold value (default threshold=50)}

\item{bg}{\link{numeric}  (\strong{with default}) background values (default bg=0)}
}
\value{
filtered_image \link{array}
}
\description{
remove-noise
remove noise by deleting outliers
}
\examples{
\dontrun{
ls<-list.files("images/",pattern = "tif$",full.names = TRUE)
im<-terra::rast(ls)
im\%>\%
  as.array\%>\%
    remove_noise(radius=5,bg=505)\%>\%
      rast
}
}
