% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rij-package.R
\docType{package}
\name{Rij-package}
\alias{Rij}
\alias{Rij-package}
\title{Rij: R tools for image analysis}
\description{
some basic tools for image processing and analysis in R.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/Zink-Antoine/Rij/issues}
}

}
\author{
\strong{Maintainer}: Antoine Zink \email{antoine.zink@culture.gouv.fr} (\href{https://orcid.org/0000-0001-7146-1101}{ORCID})

}
\keyword{internal}
