% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z.axis_profil.R
\name{Z.axis_profil}
\alias{Z.axis_profil}
\title{Z.axis_profil
profile following temporal (z) axis for selected ROI}
\usage{
Z.axis_profil(im, x = seq(1, 512), y = seq(1, 512))
}
\arguments{
\item{im}{\link{array} (\strong{optional}) an image as matrix or array}

\item{x}{\link{numeric} (\strong{with default}) pixel along x-axis (default x=seq(1,512))}

\item{y}{\link{numeric}  (\strong{with default}) pixel along y-axis (default y=seq(1,512))}
}
\value{
tmp \link{array}
}
\description{
Z.axis_profil
profile following temporal (z) axis for selected ROI
}
\examples{
 \dontrun{
Z<-Z.axis_profil(im,x=seq(50,75),seq(310,320))
temp<-seq(25,445,5)
plot(temp,Z,type="l",col="blue",xlab="Temperature (°C)",ylab="intensite")
}
}
