% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_z_project.R
\name{grouped_z_project}
\alias{grouped_z_project}
\title{grouped_z_project
merged images}
\usage{
grouped_z_project(im, grouped_size, FUN = sum)
}
\arguments{
\item{grouped_size}{\link{numeric} (\strong{with default}) number of image merged in a group}

\item{FUN}{\link{function}  (\strong{with default})  function between the merged images (default FUN=sum)}
}
\value{
tmp, n \link{list}
}
\description{
grouped_z_project
merged images
}
\examples{
\dontrun{
ls<-list.files("images/",pattern = "tif$",full.names = TRUE)
im<-terra::rast(ls)
im\%>\%
  as.array\%>\%
    remove_noise(radius=5,bg=505)\%>\%
      rast
}
}
