% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-noise.R
\name{local_median}
\alias{local_median}
\title{local_median
calcul of median with size radius x radius}
\usage{
local_median(im, radius, bg = 0)
}
\arguments{
\item{im}{\link{array} (\strong{optional}) an image as matrix or array}

\item{radius}{\link{numeric} (\strong{with default}) radius value (default radius=5)}

\item{bg}{\link{numeric}  (\strong{with default}) background values (default bg=0)}
}
\value{
med_mat
}
\description{
local_median
calcul of median with size radius x radius
}
